refactor: Split Radio Paradise provider
